refactor: migrate ShopPages and ShopsPage to new utils
